Add downloadable quiz results and expand question bank
Implemented a feature to allow users to download their quiz results as a text file after completing an assessment. Expanded the question bank in entrence-quiz.js with additional questions for all grade levels. Also updated a Pre-K assessment link in Level/a.php to point to the correct HTML file.